Refactored create account POST
Switched the create account POST request to follow the example of the
profile edit page. The form is responsible for POSTing the data and
Javascript is used to simply check the validity of the form data. This
lets us avoid XHR’s and JSON, where it doesn’t really make sense.
Additionally:
 - Makes emails optional when signing up
 - Updated the valid-email regular expressions to reflect above
 - Successful account creation now redirects to the new profile instead
of the home page
 - Changed first name field to automatically be filled to username if
not given
 - Changed how empty inputs are checked, username in particular is now
checked by a multipurpose regular expression that catches empty and
invalid usernames