Merge user_profile and edit
This has been tested with Django locally, but not extensively, so it
could have bugs, just not obvious ones.
This removes the edit.html file and moves the relevant code into the
user_profile.html file, then uses a whole bunch of Django template if
statements to update depending whether we’re editing or just viewing.
This has a few changes to the CSS to ensure that both states look good,
and a change to views to render both of them.