Make JSTinker code work with Django
I won’t go everything this commit does, but basically when you request
a program the server grabs the program’s data from storage, converts it
to a JSON string which is inserted into the HTML page, where Javascript
takes it and inserts it into the editor. Save button grabs data from
the page, makes a JSON string, and sends it to the server.