Add secure execution environment
This prevents user code from being able to take control of a user’s
account. Apache is set up to serve the ourjseditor/exec director as
exec.ourjseditor.com. This sets up an iframe to contain the user’s code
on a different sub-domain. Cross-origin security restrictions in the
browser keep the code from being able to escape the iframe.