Add textbox for changing program title
There are some advantages to having a textbox rather than a
content-editable div, like allowing us to specify a maxlength. If you
can edit the program, clicking on the title div will now swap it out
for a input element, then change it back to a div after you’re done
typing.