Fix sidebar height
The sidebar had a height of 100% but started under the nav-bar, meaning
it extended down too far and created some weird white space at the
bottom of the page. This changes it to start at the top of page, have a
height of 100%, and added a spacer div with a height equal to the
height of the nav-bar to the top of the sidebar, to push its elements
down.