Add program deleting button and pop-up
Fixes #13. Adds a button to delete the current program that only shows
up if you’re viewing one of your own programs. Opens a popup when you
click on it. A whole bunch of fancy stuff. Saves {{ csrf_token }} to a
variable so it can be used in the .js file which isn’t
template-rendered.