Removes pytz from requirements; turns off translation
Recent versions of Django come with pytz as a dependency that gets
automatically installed, so it isn’t necessary to specify it in the
requirements file. Translation is a lot of work, and not a priority, so
this removes LANGUAGE_CODE and disables USE_I18N. Localization could be
dealt with on a per-page basis, and isn’t a priority, so it’s turned
off. We don’t need to deal with time zone aware dates (USE_TZ),
everything should be stored in UTC (TIME_ZONE).