Python in an old migration broke
Apparently this python was doesn’t work with newer changes to the
models.py file of user_profile. And Django can’t just create a new
database with the most recent configuration, it has to start at the
beginning and run every migration. This is change is slower, but should
work more reliably.