Create snippet to expose userData to client
This creates a template, eventually to be used for a top banner, but
for now only including a bit of Javascript which creates a userData
object, and fills it in with data using Django template syntax. It adds
template code to fetch this snippet and include it at the top of the
program Pagel though it will be easy to add this to any other page and
be able to access userData there.