Comment deletion in one big hecking commit
Fix an API bug where deleting a comment wouldn’t decrease the
reply_count of its parent. Renamed the “replies” td to “lowerRowLeft,”
because now the delete button is there too. Move a bit of code. Switch
from setting an href of # to using CSS to achieve a pointer on
hovering. Added a delete button on your own comments, which opens a
popup with a delete and a cancel button. Wrote CSS for them. Hopefully
there aren’t too many bugs.