Client-side voting
Add backend code to include “hasVoted” in the data sent to the client,
which stores whether the user has voted on the current program already.
Add an HTML table with the three vote types and buttons to vote up a
program.
Add client side Javascript to make the correct POST or DELETE requests
in order to make or unmake a vote on a program.