Add voting (#48)
* Add vote app, model, and some api

Adds an app for voting. Creates a model/migration for it. Adds an api for posting or deleting votes on programs.

* Tweaks vote API
Adds a 403 error for attempting to duplicate a vote instead of silently
returning. Adds if statements for readablity.

* Switch to a master vote_types list, add votes to program API

Instead of recreating a list every time I reference the three types of votes, there is a vote_types variable that gets imported and used in multiple places. Adds voting data to the program GET endpoint.

* Add votes to programData

Vote data is now sent to the client as part of programData when a program is opened.

* Client-side voting

Add backend code to include “hasVoted” in the data sent to the client,
which stores whether the user has voted on the current program already.
Add an HTML table with the three vote types and buttons to vote up a
program.
Add client side Javascript to make the correct POST or DELETE requests
in order to make or unmake a vote on a program.

* Change "pretty" vote type to "artistic"

"Artistic" has a broader connotation and better fits with the tone of
the other two adjectives.

* Add vote count attributes to program

Program previously had no knowlege that they had votes, votes acted independently and were counted when needed. Now program have three attributes, one for each type of vote, which incriment as votes are cast. Vote objects still exist and are used to track voting on a per user basis. This adds migrations and switches the API and program view to use the properties instead of counting votes.

* Add vote counts to user profiles

Still looks ugly, can probably improve.

* Sort programs on profile by votes

Sorts the list of programs on a user profile by the combined vote score for the program.

* Add program list page

Not quite a hot list, but a new and top list. Add All Programs link to main page. Creates an API endpoint for viewing a list of programs, sorted. Mirrors the program list view. Both on the API and the page for the master program list, limit the list to 20 items. Currently not changeable or tweak-able, nor is it possible to get more items past that.

* Add program to_dict method

Adds a method to program objects which converts them to dicts, and switches to use that method when displaying a program for the API or for a program page.

* Add initial votes as a dependecy for votes on programs

Before, attempting to migrate from a state before votes would run this migration before adding votes, causes the python to fail. This way, votes have to be migrated first.