Comment fixes
Fixes two issues:
If the last comment on the page was folded or deleted, is would  check for the next element, not find anything, and error. (Introducted by c3bd3f93bdfb962ccf592618f5fcbfe00d78c4dd, before the new comment box was there.)
If two child comments were posted on the same comment without reloading the page, the parent comment variable would be changed the first time and error the second time. This fixes that.