Adds comment editing (and other tweaks)
Adds comment editing. A new "Edit" button, a Save button and a Cancel button, and a lot of CSS to get them were they needed to be. (http://alistapart.com/article/holygrail helped.) Changes the new comment post button to have a margin on both sides, instead of a margin on the right of the textbox. Adds a min-height to the textbox. Removes a spacer element and just included a margin on the element it was next to. Fixes #51, and closes #40.