Add autocomplete and required attributes to login page
The MDN said this was a good idea to help password managers.
Also adds the "required" attributes to required fields. This means that on browers that support this (all modern browers), our client side verification for this same stuff doesn't get run. I don't think this is a bad thing.