Add support for redirects after logging in
Adds a check for ?next attribute when logging in, which should redirect you after logging in. The page already supported this if you were already logged in.
Also adds support for the parameter to the top user dropdown. So if you're on a page and log in, you'll be sent back there.