Change notifications to update preview from comment
Instead of saving the preview data for the notification (the first part of the comment) as soon as the comment was made (which means that the notification has no way of updating if the comment is deleted or edited), the notification stores a ForeignKey to the comment. The notification is deleted if the comment is, and the notification only requests the text of the comment when it itself is requested by the client.