Add Notification UI
Adds HTML and CSS to dynamicly generate a list of notifications for the current user and populate them in the notif sidebar.
Read and unread support still needs to be added.
Also changes the `notifs_as_json` tag to return escaped text, ready for insertion into JS.
Also makes a change to escape text that's inserted into the description of notifications.