Mark notifications as read when opened
Adds code to send a request to mark notifications as read when they are clicked on.
Adds a check to make sure child comments are scrolled into view if a link to them is click on the same program page.
Fixes the font size of notifacation end message on pages.
Removes some debug code.