Notifications (#59)
Add model and database code in a new "notifications" app.
The notification stores a ForeignKey to the comment. The notification is deleted if the comment is, and the notification only requests the text of the comment when it itself is requested by the client.
Creates new notifications when a comment is created.
Adds to_dict method to notifications and send them to the client, very similar to how program comments are handled.
Adds a Notification UI.
Adds HTML and CSS to dynamically generate a list of notifications for the current user and populate them in the notif sidebar.
Adds support for marking notifs as read
Adds an id to notifications.
Adds an API endpoint at /api/notifs/ID, which currently only takes PATCH requests to mark the notif as read or unread
Adds JS to the read/unread indicator to make that request ^ when clicked
Mark All as Read
Adds a Mark All Read button, with JS that sends a request for each unread notification.
Also fixes a bug where @api.standardAPIErrors wouldn't catch keyErrors on PATCH requests.
Update program.js to scroll to child comments correctly (If you pass #comment-ID, where ID is the id of a comment with a parent, the page will find the parent, expand its children, and then scroll to the child comment you requested. Important for notifications.)
Mark notifications as read when opened. (Adds code to send a request to mark notifications as read when they are clicked on.)
Adds a check to make sure child comments are scrolled into view if a link to them is click on the same program page.
Adds some more notification API endpoints. (One for getting all notifications and one for getting the number of unread notifications.)
Update notification count dynamically. (Updates the notification count in the title of the page, in the upper right, and on the userData object.)