Add editor settings
Brings in the spade code to handle editor settings.
Adds a couple of themes
Makes some changes to the order of script loading.
Removes some of the code that used to handle setting up the ace editors.
Adds the small amount of CSS needed for the settings panel
Removes the `editor.$blockScrolling = Infinity;`, since it's not needed in Ace 1.4.2