Add userWorker setting
useWorker decides whether ace parses your code for syntax errors. It's been on by default before and it's kind of annoying. It's now a toggleable option, off by default.
Also orders the options in the DEFAULT_SETTINGS object to match everywhere else.
Adds a couple default settings to DEFAULT_SETTINGS that aren't implimented in the UI.