UI for subscribing
Adds a subscribe button when looking at a profile that's not yours. Clicking it subscribes or unsubscribes you.
Does some terrible, awful things with Django templates and Javascript; stuff that will need to be re-visted with #62 and a visual refresh of this page.
Also fixes a bug with profile API endpoints. When they were fetched by USERID and the id had a -, a 404 would be thrown.