Adds backend for program publishing
Adds a source_program to comments, and makes the source_comment field nullable. This way a notification can have a comment or a program that it points to.
Adds a published_message field and a last_published_date field to programs.
Adds API code to set and get the publishedMessage of the program, and update the last_published_date field to the current time when published_message is updated.