Subscriptions and publishing (#77)
* Models and API for subscriptions
Adds a many-to-many field to user_profile to handle a list of subscriptions.
Adds an API endpoint at /api/user/USERID/subscribed, which returns if you're subscribed to the passed user, on a GET request, and can be used with a PATCH request to subscribe or unsubscribe.

* UI for subscribing
Adds a subscribe button when looking at a profile that's not yours. The use of Django templates and Javascript will need to be revisited, it's inconsistent with the rest of the site.
Closes #67.

Fixes a bug with profile API endpoints. When they were fetched by USERID and the id had a -, a 404 would be thrown.

* Adds backend for program publishing
Adds a source_program to comments, and makes the source_comment field nullable. This way a notification can have a comment or a program that it points to.
Adds a published_message field and a last_published field to programs.
Adds API code to set and get the publishedMessage of the program, and update the last_published_ field to the current time when published_message is updated.

* Display created and published date on programs
Add a display for the created and lastPublished date in the program sidebar. Adjusts "dateToString" to return "Never" if passed an invalid date (which shouldn't happen).
Also closes #27.

* Program Publishing UI
Add UI to publish programs, including a confirm pop-up with a textbox to enter a message
Refactor some of the program deletion confirm pop-up so they can share some code (also because the delete pop-up was doing terrible things like "href='javascript:")
Also fix a couple of code style things related to program deletion (spacing, semicolons, etc).