Move program title check to user-info
The title on program pages, with the title of the program, was being overwritten by the notification code. To fix this, the notification code now looks for programData.title and uses that instead of pageTitle (which isn't correct on programs).
Also removes the code that set the title on programs originally, to avoid future similar issues.
Also fixes a small bug in the previous commit where the Save button would show up on programs that aren't yours.