Fullscreen option for programs
Closes #86. Adds a fullscreen program view at /program/PRO_ID/fullscreen
Refactor the process for removing elements from the new program page. Rename the id of the program-author link element to be a little more consistent.
Fixes a CSS issue with there being too much padding over the program title.
Adds a (subjectively) nice color to links in the sidebar.
Fixes an error when publishing a program that hadn't been published before. (It tried to update the date but the date had been removed from the page.)