Default options (#108)
* Remove runningLocal

I think this is mentioned in #62. The original idea was to have the program page work, to some extent, if served as static content (with no Django). I don't think it makes sense to maintain that in this repository though, as this page has become increasingly dynamic.

Adds JSX code for the template selection.
Creates a page for selecting a template, with a python view, and updates urls.py accordingly.
Creates a Django template that loads React.

* Rename and refactor things

Rename "new" programs to "unsaved" programs. When you're looking at the editor, but haven't saved the program, that's an "unsaved" program. The "new program" page is the page where you're picking a template for a new program.
Copy some of login.css to new-program.css
Rename default.json to templates.json

* Basic functionality

Add python to grab the program template from the query string and pass it the unsaved program page.
Fix a bug where the id "template-selector" showed in both Preact and the wrapping HTML, leading to it being duplicated.

* Adds CSS to style the boxes representing each option.

* Update JSON for default programs

-Parse program templates at server start instead of every page request
-templates.json to includes escaped line breaks inside of code strings (to make diffing future changes easier)

* Fix beautify issues

-Changes JS beautify settings to better match the style used in example programs
-Replaces old beautify files with updated minified ones

Program templates are now a list. This makes it a bit harder to find the template with the requested key in Python, but is overall an improvement. It also allows us to specify the order the templates are presented, which is important.

I want the 5p.js example to be the first thing new people see.

* Adds 5 program templates

-Also adds more margin on the top of the new program page, so the content box doesn't get clipped by the user nav box.

* Remove Fork button from unsaved program page

-Fix the intenting
-Refactor, getting rid of nested if statements

* Update New Program url on home page