Refactor programs lists from Database
Write a get_programs function that can get programs out of the database and sort them very nicely. And it all uses QuerySet methods so it translates into SQL and we don't load more into memory than we need to.
Impliments this function on the main program list page, the main program list api page, and the user profile api list page. Will in the future be expanded to other places as things are rewritten.
Closes #111